Why, oh why, does sbt not give me compiler errors...

+++ worker/src/main/scala/com/overviewdocs/background/reindex/Reindexer.scala
@@ -25,7 +25,7 @@ trait Reindexer {
 class DbLuceneReindexer(
   indexClient: LuceneIndexClient,
   nBufferBytes: Long = 1024 * 1024 * 50,
-  nDocumentsPerDbFetch: Long = 300,
+  nDocumentsPerDbFetch: Int = 300,
   writeInterval: FiniteDuration = FiniteDuration(5, TimeUnit.SECONDS)
 ) extends Reindexer with HasDatabase {
   import database.api._
